about Jacks, read me
Jack Russell Network is a new crypto project built on the Polygon Network that will appeal to anyone interested in crypto games and wants to receive dividends from every transaction. Jack Russell Network offers up to 8% dividends on every transaction, making it a very attractive crypto project for investors.

In addition to dividends, Jack Russell Network has another important element that sets it apart from other crypto projects - it uses artificial intelligence to continuously improve the cryptocurrency. This means that Jack Russell Network is not only one of the most interesting projects currently, but also one of the most innovative.

The project is currently in its early stages, so it has a relatively low capitalization and a small team and community, but the advantage is that it is very secure. Jack Russell Network is a project that gains investors' trust by providing secure and reliable crypto solutions.

Overall, Jack Russell Network offers investors a very interesting opportunity. If you are looking for a new crypto project with high dividends, artificial intelligence, and security, you should definitely check out Jack Russell Network. This project offers a lot of potential and could become a major player in the crypto world in the future.

Additionally, it should be noted that Jack Russell Network also offers dance-themed NFT crypto, which is fun and could become trendy once its game is launched on the app store.

It's important to note that Jack Russell Network is a deflationary coin whose supply is constantly decreasing. At launch, 50% of the coins were burned, and no further minting is possible since we have renounced ownership. This allows for gradual reduction of the coin supply and can lead to increased scarcity, potentially driving up the value of the coin over time.

In addition to this deflationary aspect, Jack Russell Network has plans to incorporate play-to-earn features, with 5-10% of the profits being put back into the game. There are a variety of games planned, including casino games like blackjack, player-vs-player games, and the main game, Inkubator.

With these exciting features, Jack Russell Network is poised to become a major player in the crypto gaming world. Its deflationary model and play-to-earn features make it a unique and compelling investment opportunity. If you're interested in crypto gaming and want to be a part of a growing community, Jack Russell Network is definitely worth checking out.